==> sierra-analyses/1.sierra_all_datawrangling_viz.Rmd <==
---
title: "Flam Data Sierra Project"
author: "Indra Boving"
date: "4/15/2021"
output: html_document
---
#Setup:
```{r}
knitr::opts_chunk$set(echo = TRUE)
#lots of extra here, but oh well___ 
library(ggplot2)
library(gapminder)
library(tidyverse)
library(lme4)
library(here)
library(segmented)
library(nlme)
#citation('nlme')
library(lubridate)
library(plotrix)
filter = dplyr::filter 
here = here::here
select = dplyr::select
library(janitor)
#source(here("source-code", "source-segmented_lme_R"))
library(naniar)

source(here::here("sierra-analyses", "scripts_functions", "plot_facet_flam.R"))
source(here::here("sierra-analyses", "scripts_functions", "figure_info_sierra_flammability.R"))
```

# Data Wrangling: 

#### Flammability: 

```{r}
flam_seki_raw <- read_csv(here('sierra-analyses', "processed-data", "seki_flam_data_all.csv")) %>% 
  clean_names() %>% 
  filter(model == "HP") %>% 
  mutate(location = "seki", 
         ignition_num = ignition) %>% 
  select(-ignition_no_double)

unique(flam_seki_raw$ignition_num)

flam_czo_2020_raw <- read_csv(here('sierra-analyses', "processed-data", "czo_2020_flam_data_all.csv")) %>% 
  clean_names() %>% 
  filter(model == "HP") %>% 
  mutate(location = "czo") %>% 
  select(-fh_nas_imputed)

unique(flam_czo_2020_raw$ignition_num)

flam_czo_2021_raw <- read_csv(here('sierra-analyses', "processed-data", "czo_2021_flam_data_all.csv")) %>% 
  clean_names() %>% 
  filter(model == "HP") %>% 
  mutate(location = "czo", 
         lower_temp_max = surface_temp_max_within_5_secs_of_flame_end, 
         sample_length = as.double(sample_length)) %>% 
  select( -gr, -lfm, -fh_old, -surface_temp_max_within_5_secs_of_flame_end, -total) %>% 
  mutate(ignition = as.numeric(ignition))

unique(flam_czo_2021_raw$ignition_num)

flam_sierra_raw_no2021 <- bind_rows(flam_seki_raw, flam_czo_2020_raw) %>% 
  mutate(taller_temp_start = temp_start_higher, 
        taller_temp_max = temp_max_higher, 
        lower_temp_max  = temp_max, 
       # = temp_change, 
       surface_temp_start = start_temp, 
       temp_at_ignition_taller = ignition_temp_higher, 
       temp_at_ignition_lower = ignition_temp, 
       ignition_num = as.numeric(ignition), 
       prop_new_lfm = NA_real_
  ) %>% 
  mutate(notes_flam = video_notes_57, 
         .keep = "unused") %>% 
  select(-precip_2month, -x1, -gr, -lfm, -video_notes_58,
         -temp_max_higher, - temp_max, -start_temp, -ignition_temp_higher, - ignition_temp)

flam_sierra_df_noswc <- bind_rows(flam_sierra_raw_no2021, flam_czo_2021_raw) %>% 
  mutate(lfm = lfm_n_as_imputed) %>% 
  select(-lfm_n_as_imputed) %>% 
  mutate_if(is.character, str_to_lower) %>% 
   mutate(id = row_number(), 
          gww_gdw = water_wt/dry_wt, 
          pos_mpa = mpa, 
          rwc_max_wc = rwc, #these are using the max values, not the extrapolated values
          rwd_max_wc = rwd
          )
# %>% 
#     dplyr::mutate(tlp = dplyr::case_when(
#                   spp %in% c("abco") ~ "2.27", 
#                   #spp %in%c("adfa")  ~ "3.31", 
#                   spp %in% c("arpa") ~ "2.21", 
#                   spp %in% c("cade") ~ "2.32",
#                   spp %in% c("ceco") ~ "2.64", 
#                   spp %in% c("pije") ~ "2.59", 
#                   spp %in% c("quke") ~ "2.87", 
#                   TRUE ~ NA_real_
#                 )) %>% 
#     mutate(tlp = as.numeric(tlp))
  

flam_swc_df <- flam_sierra_df_noswc %>% 
  select(id, individual, gww_gdw, pos_mpa) %>% 
  filter(pos_mpa > 2.4) %>% #for points above the tlp (estimated to be around ~2)
  group_by(individual) %>% 
  drop_na(gww_gdw) %>% 
  summarise(swc_sat = lm(formula = gww_gdw ~ pos_mpa)$coefficients[["(Intercept)"]]) 

flam_sierra_df <- merge(flam_swc_df, flam_sierra_df_noswc, all.x = T) %>% 
  mutate(rwc = (gww_gdw/swc_sat)*100, 
         rwd = 100 - rwc)
##Aaaaaaand viola! A very silly large dataframe of allllll of our samples. Woo! 

dlookr::diagnose(flam_sierra_df)
```

View all: 
```{r}
flam_sierra_df %>% 
  filter(ignition == 1) %>% 
  select(spp, year_month, tti, fh, fd, gd, lfm, mpa, temp_change, prop_ignite) %>%
  gather(-spp, -year_month, key = "var", value = "value") %>%
  ggplot() +
  geom_density(aes(x = value, fill = spp), 
               alpha = .4) +
  facet_wrap(~ var, scales = "free") +
  color_fill +
  labs(y = "Value", x = "Density")
```

```{r}
flam_sierra_df %>% 
  filter(ignition == 1) %>% 
  select(spp, year_month, tti, fh, fd, gd, temp_change, prop_ignite) %>%
  gather(-spp, -year_month, key = "var", value = "value") %>%
  ggplot() +
  geom_density(aes(x = value, fill = spp), 
               alpha = .4) +
  facet_wrap(~ var, scales = "free") +
  color_fill +
  labs(y = "Value", x = "Density")
```
```{r}
all_cor <- flam_sierra_df %>% 
  filter(ignition == 1) %>% 
  select(tti, fh, fd, gd, temp_change, prop_ignite, lfm, mpa) %>% 
  drop_na() %>% 
  cor()

corrplot::corrplot(all_cor, method = 'number', order = 'FPC', type = 'lower', diag = FALSE)
```


#####Visualize: 

```{r}
flam_sierra_df %>% 
  filter(ignition == 1, 
         year_month != "2021_august") %>% 
  select(spp, year_month, location, tti, fh, fd, gd, lfm, mpa, temp_change, prop_ignite) %>%
  gather(-spp, -year_month, -lfm, -location, key = "var", value = "value") %>%
  ggplot() +
  geom_point(aes(y = value, 
                 x = lfm,
                 color = spp, 
                 shape = interaction(year_month,location)), 
               alpha = .4) +
  facet_wrap(~ var, scales = "free") 
```
```{r}
plot_mpa_df <- flam_sierra_df %>% 
  filter(ignition == 1, 
        year_month != "2021_august"
         ) %>% 
  select(spp, year_month, tti, fh, fd, gd, lfm, mpa, temp_change, prop_ignite) %>%
  gather(-spp, -year_month, -mpa,  key = "flam_metric", value = "value")
  
  
for (i in unique(plot_mpa_df$spp)) {
  
 df <- plot_mpa_df %>% 
    filter(spp == i)
  
  p <- plot_facet_flam(df, mpa, value, year_month, flam_metric)+
    labs(title = i)
  
  if (!is.null(p)) plot(p)
  
}
```
Outliers: 

```{r}
quke_mpa <- flam_sierra_df %>% 
  #select(spp, individual, sample_wt, mpa, lfm) %>% 
  filter(spp == "quke") 

ceco_mpa <- flam_sierra_df  %>% 
 #select(spp, individual, sample_wt, mpa, lfm) %>% 
  filter(spp == "ceco") #not enough ignitions to do anything really --> filter out ceco?

pije_mpa <- flam_sierra_df %>% 
# select(spp, individual, sample_wt, mpa, lfm) %>% 
  filter(spp == "pije") %>% 
  replace_with_na_at("lfm", ~.x < 55) 

cade_mpa <- flam_sierra_df %>% 
 # select(spp, individual, sample_wt, mpa, lfm) %>% 
  filter(spp == "cade") %>% 
  replace_with_na_at("mpa", ~.x > 7)

arpa_mpa <- flam_sierra_df %>% 
  # select(spp, individual, sample_wt, mpa, lfm) %>% 
  filter(spp == "arpa",)%>% 
   replace_with_na_at("mpa", ~.x > 9) 

abco_mpa <- flam_sierra_df %>% 
 # select(spp, individual, sample_wt, mpa, lfm) %>% 
  filter(spp == "abco") %>% 
  replace_with_na_at("mpa", ~.x > 6) 

#combine back together: 
plot_mpa_df <- bind_rows(abco_mpa, arpa_mpa, cade_mpa, pije_mpa, ceco_mpa, quke_mpa)
```


```{r}
plot_lfm_df <- flam_sierra_df %>% 
  filter(ignition == 1, 
         year_month != "2021_august") %>% 
  select(spp, year_month, tti, fh, fd, gd, lfm, lfm, temp_change, prop_ignite) %>%
  gather(-spp, -year_month, -lfm,  key = "flam_metric", value = "value")
  
  
for (i in unique(plot_lfm_df$spp)) {
  
  df <- plot_lfm_df %>% 
    filter(spp == i)
  
  p <- plot_facet_flam(df, lfm, value, year_month, flam_metric) +
    labs(title = i)
  
  if (!is.null(p)) plot(p)
  
}
```
Deal with outliers from the above : 

```{r}
quke_lfm <- plot_mpa_df %>% 
  filter(spp == "quke") %>% 
  replace_with_na_at("lfm", ~.x > 190) 

ceco_lfm <- plot_mpa_df %>% 
  filter(spp == "ceco") #not enough ignitions to do anything really --> filter out ceco?

pije_lfm <- plot_mpa_df %>% 
  filter(spp == "pije") %>% 
  replace_with_na_at("lfm", ~.x < 55)  %>% 
  replace_with_na_at("tti", ~.x > 200) 

cade_lfm <- plot_mpa_df %>% 
  filter(spp == "cade") %>% 
  replace_with_na_at("gd", ~.x > 300)

arpa_lfm <- plot_mpa_df %>% 
  filter(spp == "arpa") %>% #remove 1 row
 replace_with_na_at("tti", ~.x > 300) 

abco_lfm <- plot_mpa_df %>% 
  filter(spp == "abco") %>% 
 # replace_with_na_at("lfm", ~.x > 6) %>% 
  replace_with_na_at("tti", ~.x > 300) 

#combine back together: 
plot_lfm_df <- bind_rows(abco_lfm, arpa_lfm, cade_lfm, pije_lfm, ceco_lfm, quke_lfm)
```

```{r}
plot_sw_df <- flam_sierra_df %>% 
  filter(ignition == 1, 
         year_month != "2021_august") %>% 
  select(spp, year_month, tti, fh, fd, gd, lfm, lfm, temp_change, prop_ignite, sample_wt) %>%
  gather(-spp, -year_month, - sample_wt,  key = "flam_metric", value = "value")
  
  
for (i in unique(plot_sw_df$spp)) {
  
  df <- plot_sw_df %>% 
    filter(spp == i)
  
  p <- plot_facet_flam(df, sample_wt, value, year_month, flam_metric) +
    labs(title = i)
  
  if (!is.null(p)) plot(p)
  
}
```

```{r}
quke_sw <- plot_lfm_df %>% 
  filter(spp == "quke") %>% 
  replace_with_na_at("sample_wt", ~.x > 5) 

ceco_sw <- plot_lfm_df %>% 
  filter(spp == "ceco") #not enough ignitions to do anything really --> filter out ceco?

pije_sw <- plot_lfm_df %>% 
  filter(spp == "pije") %>% 
  replace_with_na_at("sample_wt", ~.x > 18) 

cade_sw <- plot_lfm_df %>% 
  filter(spp == "cade") 

arpa_sw <- plot_lfm_df %>% 
  filter(spp == "arpa") 

abco_sw <- plot_lfm_df %>% 
  filter(spp == "abco") %>% 
  replace_with_na_at("sample_wt", ~.x > 9) 

#combine back together: 
flam_plot_sw_df <- bind_rows(abco_sw, arpa_sw, cade_sw, pije_sw, ceco_sw, quke_sw)
```

```{r}
plot_rwc_df <- flam_sierra_df %>% 
  filter(ignition == 1, 
         year_month != "2021_august", 
         rwc < 100, 
         rwc > 0
         ) %>% 
  select(spp, year_month, tti, fh, fd, gd, lfm, lfm, temp_change, prop_ignite, rwc) %>%
  gather(-spp, -year_month, - rwc,  key = "flam_metric", value = "value")
  
  
for (i in unique(plot_rwc_df$spp)) {
  
  df <- plot_rwc_df %>% 
    filter(spp == i)
  
  p <- plot_facet_flam(df, rwc, value, year_month, flam_metric) +
    labs(title = i)
  
  if (!is.null(p)) plot(p)
  
}
```

#Write csv
```{r}
# flam_deps_final_1 <- cbind(plot_lfm_df, plot_mpa_df
#                               #by = c("spp", "individual")
#                               )
# 
# flam_deps_final_2 <- cbind(plot_sw_df, flam_deps_final_1
#                               #by = c("spp", "individual"), all.y = T
#                            )

flam_sierra_df_1 <- flam_sierra_df %>% 
                          select(-lfm, -sample_wt, - mpa, -individual)

flam_sierra_df_2 <-  merge(flam_plot_sw_df, flam_sierra_df_1, all.y = T)
  

flam_sierra_df_2 %>% 
  write_csv(here('sierra-analyses', "processed-data", "sierra_flam_data_all.csv"))
```
