==> sierra-analyses/2.PCA.Rmd <==
---
title: "PCA Figure and Table"
author: "Indra Boving & Joe Celebrezze"
date: "6/20/2022"
output: html_document
---

OCT 29 NOTES: Probably will not use PCA - doesn't really tell as nice of a story with all species.
DEC 14 NOTES: Actually we might! 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Setup
Necessary packages:
```{r}
library(here)
library(kableExtra)
#remotes::install_github("vqv/ggbiplot")
#install.packages("ggbiplot")
library(ggbiplot) 
library(tidyverse)
library(sjPlot)
library(ggpubr)
library(remef)
library(lme4)

here = here::here
rename = dplyr::rename
select = dplyr::select

source(here::here("sierra-analyses", "scripts_functions", "figure_info_sierra_flammability.R")) #color and theme info is here
```

# 1.1. PCA Figure

Plot setup:
```{r warning=FALSE}
### Note: This dataset includes only Epiradiator and Ignited samples
sierra_flamdata_pca_all <- read_csv(here('sierra-analyses', "processed-data", "sierra_flam_data_all.csv"), show_col_types = FALSE) 

sierra_flamdata_pca <- sierra_flamdata_pca_all %>% 
  filter(ignition == 1 
        # ignition_num == 1
         ) %>% 
  mutate(F.Group = case_when(
    spp == "arpa" ~ "Angiosperm", 
    spp == "abco" ~ "Gymnosperm",
    spp == "cade" ~ "Gymnosperm",
    spp == "ceco" ~ "Angiosperm",
    spp == "pije" ~ "Gymnosperm",
    spp == "quke" ~ "Angiosperm"
  )) %>% 
   select("fh", "ttfg", "tti", "fd", "gd", "gti", "pfg","temp_change", "hydration", "spp", "year_month", "sample", "F.Group") %>% 
  na.omit(c("fh", "ttfg", "tti", "fd", "gd", "gti", "pfg","temp_change", "hydration", "spp", "year_month", "sample"))

 # filter(year == 2020)

figure_pca_quant <- (sierra_flamdata_pca[, c("fh", "ttfg", "tti", "fd", "gd", "gti", "pfg","temp_change")]) %>%  
 transmute( "               Flame Duration" = fd, #using spaces to shift labels over instead of \t since \t is leading to little blank squares popping up on the plot.. spaces are just manual tabs anyways right?
           "                 Flame Height" = fh,
           "Post-flame Glow" = pfg,
          "Time to Ignition" = tti,
         # "Max. Temp" = taller_temp_max,
            "Glow Duration" = gd,
          "\nGlow to Ignition" = gti,
          "\nTime to First Glow" = ttfg, 
          "Temp Change" = temp_change
          ) %>% 
  na.omit()

figure_pca_cat <-  (sierra_flamdata_pca[, c("fh", "ttfg", "tti", "fd", "gd", "gti", "pfg","temp_change", "hydration", "spp", "year_month", "sample", "F.Group")]) %>% 
  na.omit() %>% 
  select(hydration, spp, year_month, sample, F.Group) %>% 
  transmute("Hydration" = hydration, 
            spp = spp,
            "Species" = spp,
            "Timing" = year_month, 
            "Sample" = sample) 
  
  figure_prcomp <- prcomp(na.omit(figure_pca_quant), 
                          center = TRUE, 
                          scale = TRUE, 
                          retx=TRUE)
  
figure_pca <- (sierra_flamdata_pca[, c("fh", "ttfg", "tti", "fd", "gd", "gti", "pfg","temp_change", "hydration", "spp", "year_month", "sample", "F.Group")]) %>% 
  na.omit() 
```


```{r warning=FALSE}
#Setting up plotting data: 

#decisions for plotting:
choices = 1:2 
scale = 1
obs.scale = 1 - scale
var.scale = scale
ellipse.prob = 0.68
labels.size = 3
circle.prob = 0.69
choices = 1:2

#Run PCA
pcobj <- prcomp(na.omit(figure_pca_quant), center = TRUE, scale = TRUE)

# extract PCA components: 
nobs.factor <- sqrt(nrow(pcobj$x) - 1)
    d <- pcobj$sdev
    u <- sweep(pcobj$x, 2, 1/(d * nobs.factor), FUN = "*")
    v <- pcobj$rotation

  choices <- pmin(choices, ncol(u))
  df.u <- as.data.frame(sweep(u[, choices], 2, d[choices]^obs.scale, 
                              FUN = "*"))
  v <- sweep(v, 2, d^var.scale, FUN = "*")
  df.v <- as.data.frame(v[, choices])
  names(df.u) <- c("PC1", "PC2")
  names(df.v) <- names(df.u)
  df.u <- df.u * nobs.factor

  r <- sqrt(qchisq(circle.prob, df = 2)) * prod(colMeans(df.u^2))^(1/4)
  
v.scale <- rowSums(v^2)
df.v <- r * df.v/sqrt(max(v.scale)) 
df.v <- df.v %>% mutate(PC1, PC2)
df.v$Variables <- rownames(v) 
PCAloadings = df.v
#df.v = dataset with loadings 

  #dataset with scores and categorical variables for plotting points: 
PCAvalues <- cbind(df.u, figure_pca_cat)

#dataset with information for plotting circle: 
theta <- c(seq(-pi, pi, length = 50), seq(pi, -pi, length = 50))

circle <- data.frame(PC1 = r * cos(theta), PC2 = r * sin(theta)) 

#Group by flam. metric: 
PCA_combustability <- df.v %>% 
  filter(Variables %in% c("                 Flame Height", "Max. Temp","               Flame Duration"))
         
PCA_ignitability <- df.v %>% 
  filter(Variables %in% c("Time to Ignition", "\nGlow to Ignition"))

PCA_consumability <- df.v %>% 
  filter(Variables %in% c("Glow Duration", "Post-flame Glow", "\nTime to First Glow"))

# Calculate the angles and the label offset
df.v$Angle = ((180/pi) * atan(df.v$PC2/df.v$PC1))

df.v$Offset <- ((-2 * sign(df.v$PC1))/2)

#labels: 
u.axis.labs <- paste("PC", choices, sep = "")
u.axis.labs <- paste(u.axis.labs, sprintf("(%0.1f%% explained var.)", 
                                            100 * pcobj$sdev[choices]^2/sum(pcobj$sdev^2)))
```

Plot:
```{r}
ggplot(PCAvalues, aes(x = PC1, y = PC2)) +
  ylim(-4, 3) +
  geom_point(size = 1.75, alpha = .45, aes(shape = Species
                                       , color = Species)
                                       , data = PCAvalues) +
  guides(color = guide_legend(override.aes = list(size = 3))) +
  geom_segment(data = PCA_combustability, aes(x = 0, y = 0, 
                                              xend = PC1, yend = PC2),
     arrow = arrow(length = unit(1/2, "picas")), 
      size = .9, alpha = 0.9,
     color = "#a56457") +
  geom_segment(data = PCA_consumability, aes(x = 0, y = 0,
                                             xend = PC1, yend = PC2),
     arrow = arrow(length = unit(1/2, "picas")), 
     size = .9, alpha = 0.9,
     color = "#b08ba5") +
  geom_segment(data = PCA_ignitability, aes(x = 0, y = 0, 
                                             xend = PC1, yend = PC2),
     arrow = arrow(length = unit(1/2, "picas")), 
     size = .9, alpha = 0.9,
     color = "#ffb178") +
  annotate("text", 
           x = (PCA_combustability$PC1 * 0.7), 
          y = (PCA_combustability$PC2 * 1.15),
          label = PCA_combustability$Variables, size = 3.5,
          fontface = 'bold') +
  annotate("text",
           x = (PCA_ignitability$PC1 * 1.3), 
           y = (PCA_ignitability$PC2 * 1.45), 
           label = PCA_ignitability$Variables, size = 3.5,
           fontface = 'bold') +
  annotate("text",
           x = (PCA_consumability$PC1 * 1.3),
           y = (PCA_consumability$PC2 * 1.1), 
           label = PCA_consumability$Variables, size = 3.5,
           fontface = 'bold') + 
  geom_segment(aes(x = -3.1, xend = -3.1, y = -3.45, yend = -1.85), arrow = arrow(length = unit(1/2, 'picas')), size = 0.35, alpha = 0.85) +
  annotate('text', x = -3.25, y = -2.6, label = "Less Consumable", angle = 90, size = 3.5) +
  geom_segment(aes(x = -3.1, xend = -1.35, y = -3.45, yend = -3.45), arrow = arrow(length = unit(1/2, 'picas')), size = 0.35, alpha = 0.85) +
  annotate('text', x = -2.25, y = -3.68, label = "Larger & Hotter Flames, \n Faster Ignition", size = 3.5) +
  theme(panel.background = element_rect(fill='white', colour='black'), # Make background white and border black
          panel.grid.major = element_blank(),  # Hide major gridlines
          panel.grid.minor = element_blank(),
          legend.key = element_rect(fill = "white"),
          legend.position = c(0.2,0.9),
          legend.title = element_text(face = 'bold', size = 15),
          legend.text = element_text(face = 'italic', size = 13),
          axis.title = element_text(face = 'bold', size = 15)) +
 # scale_color_manual(values = c("#9D0208","#F48C06")) + 
  geom_path(data = circle, color = "black",  size = 1/2, alpha = 1/3) + 
  xlab(u.axis.labs[1]) + 
   ylab(u.axis.labs[2]) +
  coord_equal()

ggsave(here('sierra-analyses',"figures", "pca.jpg"),
       height = 7, width = 9)
```

## Adding principal components into dataset

```{r warning=FALSE}
x_flam <-predict(pcobj)  ### generate scores for each row in sierra_flamdata_pca

pca_flam_data <- as_data_frame(x_flam) 

all_data_pca <- bind_cols(sierra_flamdata_pca, pca_flam_data)

# #because we wants instances where they didn't ignite for the logistic regression: 
sierra_flamdata_pca_noignitions <- sierra_flamdata_pca_all %>%
 # mutate(LFM = lfm_NAs_imputed) %>%
  #filter(year == 2020) %>%
 # filter(model == "EPI") %>%
  filter(ignition == 0)

all_data_withnoignitions <- bind_rows(all_data_pca, sierra_flamdata_pca_noignitions)

#dataset with the above columns, PCA using imputed NAs (medians), and flam metrics with NAs still in there:
mem_data <- all_data_withnoignitions

write_csv(mem_data, here('sierra-analyses', "processed-data", "mem_data_sierra_alldates.csv"))
```

# 1.2. PCA Table

```{r warning=FALSE}
figure_pca_cat <- (sierra_flamdata_pca[c("hydration", "spp", "year_month", "sample")]) %>% 
  transmute("Hydration" = hydration, 
            spp = spp,
            "Timing" = year_month, 
            "Sample" = sample) %>% 
  mutate(Species = case_when(
    spp ==  "ARPA" ~ "A. patula", 
    spp == "ABCO" ~ "A. concolor", 
    spp == "CADE" ~ "C. decurrens", 
    spp == "PIJE" ~ "P. jeffryii", 
    spp == "CECO" ~ "C. cordulatus",
    spp == "QUKE" ~ "Q. kellogii"
  ))


figure_varimax <- varimax(pcobj$rotation[,1:3])
figure_varimax

sjPlot::tab_pca(pcobj, 
        rotation = c("varimax"), 
        digits = 2,
        show.var = TRUE,
  string.pov = "Proportion of Variance",
  string.cpov = "Cumulative Proportion") 

```

#---

# 2. Interspecific differences

## PCA Setup
```{r}
sierra_flamdata_pca_noceco <- sierra_flamdata_pca_all %>% 
  filter(ignition == 1
         ) %>% 
  mutate(F.Group = case_when(
    spp == "arpa" ~ "Angiosperm", 
    spp == "abco" ~ "Gymnosperm",
    spp == "cade" ~ "Gymnosperm",
    spp == "ceco" ~ "Angiosperm",
    spp == "pije" ~ "Gymnosperm",
    spp == "quke" ~ "Angiosperm"
  )) %>% 
  filter(spp != "ceco") %>% 
   select("fh", "ttfg", "tti", "fd", "gd", "gti", "pfg","lower_temp_max", "hydration", "spp", "year_month", "sample", "F.Group", "mpa", "lfm_outliers_out", "site", "individual", "sample_wt") %>% 
  na.omit(c("fh", "ttfg", "tti", "fd", "gd", "gti", "pfg","lower_temp_max", "hydration", "spp", "year_month", "sample"))
  
figure_pca_quant_spp <- (sierra_flamdata_pca_noceco[, c("fh", "ttfg", "tti", "fd", "gd", "gti", "pfg","lower_temp_max")]) %>%  
  mutate(tti = tti*-1) %>% 
  transmute("Flame Duration" = fd,
           "Flame Height" = fh,
           "Post-flame Glow" = pfg,
           "Ignitability" = tti,
           "Glow Duration" = gd,
           "Glow to Ignition" = gti,
           "Time to First Glow" = ttfg, 
           "Max Temp" = lower_temp_max
          ) %>% 
  na.omit()

figure_pca_cat_spp <-  (sierra_flamdata_pca_noceco[, c("fh", "ttfg", "tti", "fd", "gd", "gti", "pfg","lower_temp_max", "hydration", "spp", "year_month", "sample", "F.Group")]) %>% 
  na.omit() %>% 
  select(hydration, spp, year_month, sample, F.Group) %>% 
  transmute("Hydration" = hydration, 
            spp = spp,
            "Species" = spp,
            "Timing" = year_month, 
            "Sample" = sample)
  
figure_prcomp_spp <- prcomp(na.omit(figure_pca_quant_spp), 
                          center = TRUE, 
                          scale = TRUE)
  
figure_pca_spp <- (sierra_flamdata_pca_noceco[, c("fh", "ttfg", "tti", "fd", "gd", "gti", "pfg","lower_temp_max", "hydration", "spp", "year_month", "sample", "F.Group")]) %>% 
  na.omit() 

# Putting principal components into dataset
spp_pc_flam <-predict(figure_prcomp_spp) 
pca_flam_data_spp <- as_data_frame(spp_pc_flam) 
spp_data_pca <- bind_cols(sierra_flamdata_pca_noceco, pca_flam_data_spp)

spp_data_pca <- spp_data_pca %>% 
  mutate(Species = case_when(
    spp == 'abco' ~ 'Ab. concolor',
    spp == 'arpa' ~ 'Ar. patula',
    spp == 'cade' ~ 'Ca. decurrens',
    spp == 'pije' ~ 'Pi. jeffreyii',
    spp == 'quke' ~ 'Qu. kelloggii'
  ))
spp_data_pca$Species <- factor(spp_data_pca$Species, levels = c('Ab. concolor', 'Pi. jeffreyii', 'Ca. decurrens', 'Ar. patula', 'Qu. kelloggii'))
```

## PCA Visualization
### Quick and Easy
```{r}
ggbiplot(figure_prcomp_spp,
         groups = figure_pca_spp$F.Group,
         ellipse = TRUE, circle = FALSE, varname.size = 4, alpha = 0.5) +
  theme_bw() + 
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        aspect.ratio = 1/1,
        axis.title = element_text(face = 'bold'),
        legend.title = element_text(face = 'bold')) +
  labs(color = 'Functional Group')
```

### GGPlot
Setup:
```{r}
#decisions for plotting:
choices = 1:2
scale = 1
obs.scale = 1 - scale
var.scale = scale
ellipse.prob = 0.68
labels.size = 3
circle.prob = 0.69
choices = 1:2

# extract PCA components: 
nobs.factor <- sqrt(nrow(figure_prcomp_spp$x) - 1)
    d <- figure_prcomp_spp$sdev
    u <- sweep(figure_prcomp_spp$x, 2, 1/(d * nobs.factor), FUN = "*")
    v <- figure_prcomp_spp$rotation

  choices <- pmin(choices, ncol(u))
  df.u <- as.data.frame(sweep(u[, choices], 2, d[choices]^obs.scale, 
                              FUN = "*"))
  v <- sweep(v, 2, d^var.scale, FUN = "*")
  df.v <- as.data.frame(v[, choices])
  names(df.u) <- c("PC1", "PC2")
  names(df.v) <- names(df.u)
  df.u <- df.u * nobs.factor

  r <- sqrt(qchisq(circle.prob, df = 2)) * prod(colMeans(df.u^2))^(1/4)
  
v.scale <- rowSums(v^2)
df.v <- r * df.v/sqrt(max(v.scale)) 
df.v <- df.v %>% mutate(PC1, PC2)
df.v$Variables <- rownames(v) 
PCAloadings = df.v
#df.v = dataset with loadings 

# To adjust positions of labels in PCA:
df.v <- df.v %>% 
  mutate(Flam.Metric = case_when(
    Variables == "Flame Height" ~ " ", # Entered this one manually so took out the label here
    Variables == "Flame Duration" ~ "\nFlame Duration              ",
    Variables == "Glow Duration" ~ "Glow Duration",
    Variables == "Post-flame Glow" ~ "\nPost-flame Glow",
    Variables == "Ignitability" ~ "          Ignitability",
    Variables == "Glow to Ignition" ~ "Glow to Ignition",
    Variables == "Max Temp" ~ "Max Temp",
    Variables == "Time to First Glow" ~ "Time to First Glow\n",
  ))

df.v1 <- df.v %>% 
  filter(Variables %in% c("Ignitability", "Glow Duration", "Post-flame Glow",
                          "Time to First Glow"))
df.v2 <- df.v %>% 
  filter(Variables %in% c("Flame Duration", "Glow to Ignition", "Max Temp",
                          "Flame Height"))

  #dataset with scores and categorical variables for plotting points: 
PCAvalues <- cbind(df.u, figure_pca_cat_spp)

#dataset with information for plotting circle: 
theta <- c(seq(-pi, pi, length = 50), seq(pi, -pi, length = 50))

circle <- data.frame(PC1 = r * cos(theta), PC2 = r * sin(theta)) 
```

Visualization:
```{r}
ggplot(spp_data_pca, aes(x = PC1, y = PC2, color = Species)) +
  geom_point(size = 1.7, alpha = .35) +
  stat_ellipse(aes(x = PC1, y = PC2, color = Species), alpha = 0.5, cex = 0.95) +
  #labs(color = "Functional Group") +
  labs(color = "Species") +
  guides(color = guide_legend(override.aes = list(size = 3, alpha = 0.55, lty = 0))) + # For legend point appearance
  geom_segment(data = df.v, aes(x = 0, y = 0, 
                                             xend = PC1, yend = PC2),
     arrow = arrow(length = unit(1/2, "picas")), 
     size = .93, alpha = 0.9, color = 'black') +
 # scale_color_manual(values = c('steelblue3', 'darkgreen')) +
  color_noceco +
  annotate("text",
          x = df.v1$PC1*1.3,
          y = df.v1$PC2*1.15,
          label = df.v1$Flam.Metric, size = 4.25,
          fontface = 'bold') +
  annotate("text",
          x = df.v2$PC1*2.1,
          y = df.v2$PC2*1.7,
          label = df.v2$Flam.Metric, size = 4.25,
          fontface = 'bold') +
  annotate("text", x = 0.28, y = 0.55, label = "Flame Height", 
           size = 4.25, fontface = 'bold') +
  theme(panel.background = element_rect(fill='white', colour='black'), # Make background white and border black
          panel.grid.major = element_blank(),  # Hide major gridlines
          panel.grid.minor = element_blank(),
          legend.key = element_rect(fill = "white"),
          legend.title = element_text(face = 'bold', size = 18),
         # legend.text = element_text(size = 16),
        legend.text = element_text(face = 'italic', size = 16),
          axis.title = element_text(face = 'bold', size = 20),
          axis.text = element_text(size = 14),
          aspect.ratio = 1/1) +
  coord_equal()  + 
  geom_segment(
    aes(
      x = -5.5,
      xend = -5.5,
      y = -5.5,
      yend = -2
    ),
    arrow = arrow(length = unit(1 / 2, 'picas')),
    size = 0.35,
    alpha = 0.85
  ) +
  annotate(
    'text',
    x = -5.75,
    y = -3.75,
    label = "More consumable",
    angle = 90,
    size = 3.5
  ) +
  geom_segment(
    aes(
      x = -5.5,
      xend = -2,
      y = -5.5,
      yend = -5.5
    ),
    arrow = arrow(length = unit(1 / 2, 'picas')),
    size = 0.35,
    alpha = 0.85
  ) +
  annotate(
    'text',
    x = -3.75,
    y = -5.9,
    label = "More ignitable, \n Less combustable",
    size = 3.5
  ) 

#ggsave(here("figures", "PCA.angio.vs.gymno.jpg"), height = 8, width = 10)

ggsave(here("sierra-analyses", "figures", "PCA.spp.diff.jpg"), height = 8, width = 10) 

# TO RUN SPP PCA: change Species > Species, legend.text = element_text(face = italic) and change color palette to color_many2
```

## PCA Table
```{r}
tab_pca(figure_prcomp_spp, 
        rotation = c("varimax"), 
        digits = 2,
        show.var = TRUE,
  string.pov = "Proportion of Variance",
  string.cpov = "Cumulative Proportion")
```

```{r}
eig1 <- c(-0.96, 0.85, 0.00, 0.04, -0.03, -0.07, -0.09, -0.06, 26.20, 26.20)
eig2 <- c(0.08, 0.04, -0.97, -0.96, -0.16, -0.05, -0.04, 0.26, 24.74, 50.94)
eig3 <- c(-0.17, -0.45, 0.16, -0.12, 0.93, -0.05, 0.18, 0.47 ,14.76, 65.70)
eig4 <- c(-0.10, 0.12, -0.12, 0.02, 0.03, -0.77, -0.77, -0.29, 13.26, 78.96)
flam.metrics <- c("Time to First Glow", "Ignitability", "Glow Duration","Post-flame Glow", "Glow to Ignition", "Flame Duration", "Maximum Temperature", "Flame Height", "Proportion of Variance (%)", "Cumulative Proportion (%)")
pca.table.df <- data.frame(flam.metrics, eig1, eig2, eig3, eig4)
```

there was probably a more elegant way to do that, but since there aren't too many numbers to enter in by hand, I thought it would be quicker to just go ahead and do so.

Now, using kableExtra package to visualize how each component is weighted by each flam. metric

```{r}
pca.table.df %>%
  mutate(flam.metrics = cell_spec(flam.metrics, 'html', bold = ifelse(flam.metrics == c("Proportion of Variance (%)", "Cumulative Proportion (%)"), T, F), color = 'black')) %>% 
  mutate(eig1 = cell_spec(eig1, 'html', bold = ifelse(abs(eig1) > 0.5, T, F), color = 'black')) %>% 
  mutate(eig2 = cell_spec(eig2, 'html', bold = ifelse(abs(eig2) > 0.5, T, F), color = 'black')) %>% 
  mutate(eig3 = cell_spec(eig3, 'html', bold = ifelse(abs(eig3) > 0.5, T, F), color = 'black')) %>% 
  mutate(eig4 = cell_spec(eig4, 'html', bold = ifelse(abs(eig4) > 0.5, T, F), color = 'black')) %>% 
  kable(format = 'html', escape = F, col.names = c(' ', 'Component 1', 'Component 2', 'Component 3', 'Component 4')) %>% 
  kable_styling(bootstrap_options = c('hover', 'bordered', 'condensed'), fixed_thead = T, font_size = 30) %>% 
  row_spec(c(9:10), background = '#D3D3D3') %>% 
  row_spec(0, bold = T, color = 'black') %>% 
  #save_kable(here('figures', 'pca_table.html'))
save_kable(here('sierra-analyses','figures', 'pca_table.jpeg'))

```


## PC's vs. LFM
### PC1
```{r}
# MEM and REMEF
pc1_mod <- lmer(PC1 ~ spp*mpa + spp*lfm_outliers_out + site + year_month + sample_wt + (1 | individual), data = spp_data_pca)
r_pc1_lfm <- remef(pc1_mod, fix = c("mpa", "sample_wt", "site"), ran = list(individual = c("(Intercept)"))) # Using Remef for plots with LFM
spp_data_pca$r_pc1_lfm <- r_pc1_lfm

# PC1 vs. LFM visualization
pc1_vs_lfm_r <- spp_data_pca %>%
  ggplot(aes(y = r_pc1_lfm, 
             x = lfm_outliers_out, 
             group = F.Group)) +
  geom_point(alpha= .4, aes(color = F.Group))+
  geom_smooth(method = "lm", se = F, aes(color = F.Group), alpha = .75) + 
  scale_color_manual(values = c('#FDD358', 'black')) +
  theme(axis.text.y = element_text(size = 13),
        axis.title = element_text(size = 16, face = 'bold')) +
        #axis.text.x = element_blank(), axis.ticks.x = element_blank()) +
  labs(x = "Live Fuel Moisture (%)", 
       y = "PC1",
       color = "Functional Group") +
  annotate('text', x = 35, y = -4.5, label = 'More Combustible, \nMore Sustainable', size = 3) +
  geom_segment(aes(x = 35, xend = 35, y = -4.2, yend = -2.8), arrow = arrow(length = unit(1/2, 'picas')), size = 0.35, alpha = 0.85) +
  annotate('text', x = 35, y = -2.5, label = 'More Ignitable', size = 3)
pc1_vs_lfm_r
```

### PC2
```{r}
# MEM and REMEF
pc2_mod <- lmer(PC2 ~ spp*mpa + spp*lfm_outliers_out + site + year_month + sample_wt + (1 | individual), data = spp_data_pca)
r_pc2_lfm <- remef(pc2_mod, fix = c("mpa", "sample_wt", "site"), ran = list(individual = c("(Intercept)"))) # Using Remef for plots with LFM
spp_data_pca$r_pc2_lfm <- r_pc2_lfm

# PC2 vs. LFM visualization
pc2_vs_lfm_r <- spp_data_pca %>%
  ggplot(aes(y = r_pc2_lfm, 
             x = lfm_outliers_out, 
             group = F.Group)) +
  geom_point(alpha= .4, aes(color = F.Group))+
  geom_smooth(method = "lm", se = F, aes(color = F.Group), alpha = .75) + 
  scale_color_manual(values = c('#FDD358', 'black')) +
  theme(axis.text.y = element_text(size = 13),
        axis.title = element_text(size = 16, face = 'bold'),
        axis.text.x = element_blank(), axis.ticks.x = element_blank()) +
  labs(x = "Live Fuel Moisture (%)", 
       y = "PC2",
       color = "Functional Group") +
  annotate('text', x = 35, y = 4.5, label = 'More Consumable', size = 3) +
  geom_segment(aes(x = 35, xend = 35, y = 2.8, yend = 4.2), arrow = arrow(length = unit(1/2, 'picas')), size = 0.35, alpha = 0.85)
pc2_vs_lfm_r
```
